imp: userModel all in one, use discriminatorKey - role
