--- backend/src/models/producerModel.ts
@@ -1,35 +1,0 @@
-import mongoose, { Document, Schema } from "mongoose";
-import { IUser, userSchema } from "./userModel";
-
-// Define the interface for Producer, extending IUser
-export interface IProducer extends IUser {
-  company?: string;
-  paymentType?: "qrCode" | "creditDebit";
-  nameOnCard?: string; //for Credit/Debit
-  cardNumber?: string; //for Credit/Debit
-}
-
-// Define the schema for Producer, inheriting from userSchema
-export const producerSchema = new Schema<IProducer>({
-  company: {
-    type: String,
-    trim: true,
-  },
-  paymentType: {
-    type: String,
-    enum: ["qrCode", "creditDebit"],
-  },
-  nameOnCard: {
-    type: String,
-  },
-  cardNumber: {
-    type: String,
-  },
-});
-
-// Attach the userSchema as the base schema
-producerSchema.add(userSchema);
-
-// Export the model
-const Producer = mongoose.model<IProducer>("producer", producerSchema, "producers");
-export default Producer;
